chore: pass platform config through compose and clean clippy warnings
docker-compose passes OMON_PLATFORM and SLACK_* to the container with a
clear error when DEFAULT_MODEL is unset. Fixes all clippy warnings:
sort_by_key with Reverse, redundant cast and static lifetime, unused
import, tokio Mutex for env-guarding tests, expect_err.

+++ src/slack/runtime.rs
@@ -422,7 +422,7 @@ impl SlackRuntime {
             };
 
             let small_enough = size_bytes.is_some_and(|size| {
-                size > 0 && size <= MAX_INLINED_ATTACHMENT_BYTES as u64
+                size > 0 && size <= MAX_INLINED_ATTACHMENT_BYTES
             });
             if small_enough
                 && !attachment.url.is_empty()
